requirements.txt
streamlit
pandas
tensorflow
opencv-python
fer
facenet-pytorch